+++ src/definitions.ts
@@ -107,6 +107,28 @@ export interface updateAvailableEvent {
   bundle: BundleInfo;
 }
 
+export interface channelRes {
+  /**
+   * Current status of set channel
+   *
+   * @since  4.7.0
+   */
+  status: string;
+  error?: any;
+}
+
+export interface getChannelRes {
+  /**
+   * Current status of get channel
+   *
+   * @since  4.8.0
+   */
+  channel?: string;
+  error?: any;
+  status?: string,
+  allowSet?: boolean,
+}
+
 export interface DownloadEvent {
   /**
    * Current status of download, between 0 and 100.
@@ -171,6 +193,13 @@ export interface BundleInfo {
   status: BundleStatus;
 }
 
+export interface SetChannelOptions {
+  chhanel: string;
+}
+
+export interface SetCustomIdOptions {
+  chhanel: string;
+}
 export interface DelayCondition {
   /**
    * Set up delay conditions in setMultiDelay
@@ -322,6 +351,35 @@ export interface CapacitorUpdaterPlugin {
    * @since 4.0.0
    */
   getLatest(): Promise<latestVersion>;
+
+  /**
+   * Set Channel for this device
+   *
+   * @returns {Promise<channelRes>} an Promise resolved when channel is set
+   * @param options is the {@link SetChannelOptions} channel to set
+   * @throws An error if the something went wrong
+   * @since 4.7.0
+   */
+  setChannel(options: SetChannelOptions): Promise<channelRes>;
+
+    /**
+   * get Channel for this device
+   *
+   * @returns {Promise<channelRes>} an Promise resolved with channel info
+   * @throws An error if the something went wrong
+   * @since 4.8.0
+   */
+  getChannel(): Promise<getChannelRes>;
+
+    /**
+   * Set Channel for this device
+   *
+   * @returns {Promise<void>} an Promise resolved instantly
+   * @param options is the {@link SetCustomIdOptions} customId to set
+   * @throws An error if the something went wrong
+   * @since 4.9.0
+   */
+  setCustomId(options: SetCustomIdOptions): Promise<void>;
 
   /**
    * Listen for download event in the App, let you know when the download is started, loading and finished